Don't warn when the file being deleted does not really exist. This happens when opening the External Material dialog, where a temporary filename is generated but the corresponding file is not yet created.
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@21893 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/support/FileName.cpp
@@ -262,7 +262,7 @@ unsigned long FileName::checksum() const
 bool FileName::removeFile() const
 {
 	bool const success = QFile::remove(d->fi.absoluteFilePath());
-	if (!success)
+	if (!success && exists())
 		lyxerr << "FileName::removeFile(): Could not delete file "
 			<< *this << "." << endl;
 	return success;
